refactor(naming): prefer professional internal references
