+++ src/sdetkit/readiness_boost.py
@@ -1,9 +1,9 @@
-"""Compatibility wrapper for historical `sdetkit.phase_boost` imports."""
+"""Compatibility wrapper for historical `sdetkit.readiness_boost` imports."""
 
 from __future__ import annotations
 
 from importlib import import_module as _import_module
 
-_IMPL = _import_module("sdetkit.phases.phase_boost")
+_IMPL = _import_module("sdetkit.phases.readiness_boost")
 __all__ = getattr(_IMPL, "__all__", [name for name in dir(_IMPL) if not name.startswith("__")])
 globals().update({name: getattr(_IMPL, name) for name in __all__})
