+++ tests/test_checks_artifacts.py
@@ -29,7 +29,7 @@ def _sample_records(repo_root: Path) -> tuple[CheckRecord, ...]:
                 "truth_level": "merge",
                 "target_mode": "targeted",
                 "target_reason": "targeted pytest scope from changed files",
-                "changed_paths": ["src/sdetkit/demo.py"],
+                "changed_paths": ["src/sdetkit/example.py"],
                 "selected_targets": ["tests/test_demo.py"],
                 "cache": {"status": "hit"},
             },
@@ -76,7 +76,7 @@ def test_render_record_artifacts_writes_stable_schema_outputs(tmp_path: Path) ->
             "source": "unit-test",
             "requested_profile": "adaptive",
             "adaptive_reason": "small code change set keeps adaptive on standard validation",
-            "changed_files": ["src/sdetkit/demo.py"],
+            "changed_files": ["src/sdetkit/example.py"],
             "changed_areas": ["source", "tests"],
             "execution": {"mode": "parallel", "workers": 2},
             "checks_recorded": 3,
